fix: 사전질문 관리 메뉴 분리
